requirements.txt
amqp==5.2.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
googletrans==4.0.0rc1
greenlet==3.0.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.4.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
kombu==5.3.6
loguru==0.7.2
mysqlclient==2.2.4
pillow==10.3.0
prompt-toolkit==3.0.43
python-dateutil==2.9.0.post0
python-slugify==8.0.4
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
text-unidecode==1.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
win32-setctime==1.1.0
